requirements.txt
openai
pdfminer.six
Pillow
PyPDF2
Requests
streamlit
streamlit_option_menu
python-dotenv
setuptools
wheel